- user management WIP

+++ src/main/java/com/semestral_project/company_process_tool/services/ProcessService.java
@@ -2,10 +2,7 @@ package com.semestral_project.company_process_tool.services;
 
 import com.semestral_project.company_process_tool.entities.*;
 import com.semestral_project.company_process_tool.entities.Process;
-import com.semestral_project.company_process_tool.repositories.BPMNfileRepository;
-import com.semestral_project.company_process_tool.repositories.ElementRepository;
-import com.semestral_project.company_process_tool.repositories.ProcessMetricRepository;
-import com.semestral_project.company_process_tool.repositories.ProcessRepository;
+import com.semestral_project.company_process_tool.repositories.*;
 import com.semestral_project.company_process_tool.utils.ProcessAndBpmnHolder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -26,6 +23,14 @@ public class ProcessService {
     BPMNfileRepository bpmNfileRepository;
     @Autowired
     ProcessMetricRepository processMetricRepository;
+    @Autowired
+    UserRepository userRepository;
+    @Autowired
+    ProjectService projectService;
+    @Autowired
+    TaskService taskService;
+    @Autowired
+    TaskRepository taskRepository;
 
     public Process fillProcess(Process oldProcess, Process updatedProcess){
         oldProcess.setName(updatedProcess.getName());
@@ -61,43 +66,195 @@ public class ProcessService {
         else return null;
     }
 
-    public boolean addProcess(Process process){
+    public long addProcess(Process process, long userId){
         try {
-            processRepository.save(process);
-            return true;
-        }
-        catch (Exception e)
-        {
-            System.out.println(e.getMessage());
-            return false;
-        }
-    }
-
-    public boolean deleteProcessById(long id){
-        try {
-            if(bpmnParser.removeProcessFromAllWorkflows(processRepository.findById(id).get()))
-            {
-                processRepository.deleteById(id);
-                return true;
+            if(userRepository.existsById(userId)) {
+                User user = userRepository.findById(userId).get();
+                if(process.getProject() != null){
+                    Project project = projectService.getProjectById(process.getProject().getId());
+                    if(project.getCanEdit().contains(user)){
+                        var list = process.getCanEdit();
+                        list.add(user);
+                        process = processRepository.save(process);
+                        return process.getId();
+                    } else {
+                        return -1;
+                    }
+                }
+                var list = process.getCanEdit();
+                list.add(user);
+                process = processRepository.save(process);
+                return process.getId();
             }
-            return false;
+            else return -1;
         }
         catch (Exception e)
         {
             System.out.println(e.getMessage());
-            return false;
+            return -1;
+        }
+    }
+//
+//    public boolean addProcess(Process process){
+//        try {
+//            processRepository.save(process);
+//            return true;
+//        }
+//        catch (Exception e)
+//        {
+//            System.out.println(e.getMessage());
+//            return false;
+//        }
+//    }
+
+    public int addAccess(long processId, long whoEdits, User getAccess){
+        Optional<Process> processData = processRepository.findById(processId);
+        if(processData.isPresent()) {
+            Process process_ = processData.get();
+            User whoEdits_ = userRepository.findById(whoEdits).get();
+            if(process_.getCanEdit().contains(whoEdits_)){
+                User getAccess_ = userRepository.findById(getAccess.getId()).get();
+                if(process_.getCanEdit().contains(getAccess_)){
+                    return 4; //already can edit
+                } else if(process_.getHasAccess().contains(getAccess_)) {
+                    return 3; //already has access
+                } else{
+                    var list = process_.getHasAccess();
+                    list.add(getAccess_);
+                    process_.setHasAccess(list);
+                    processRepository.save(process_);
+                    for(Element e : process_.getElements())
+                    {
+                        if(e.getClass() == Task.class){
+                            taskService.addEdit(e.getId(), whoEdits_.getId(),getAccess);
+                        } else {
+                            this.addEdit(e.getId(),whoEdits_.getId(), getAccess); //
+                        }
+                    }
+                    return 1; //OK
+                }
+            }else return 5; //cannot edit
+        }
+        else
+        {
+            return 2; //role not found
         }
     }
 
-    public int updateProcess(long id, Process process){
-        Optional<Process> processData = processRepository.findById(id);
+    public int removeAccess(long processId, long whoEdits, User removeAccess){
+        Optional<Process> processData = processRepository.findById(processId);
+        if(processData.isPresent()) {
+            Process process_ = processData.get();
+            User whoEdits_ = userRepository.findById(whoEdits).get();
+            if(process_.getCanEdit().contains(whoEdits_)){
+                User getAccess_ = userRepository.findById(removeAccess.getId()).get();
+                if(process_.getHasAccess().contains(getAccess_)) {
+                    var list = process_.getHasAccess();
+                    list.remove(getAccess_);
+                    process_.setHasAccess(list);
+                    processRepository.save(process_);
+                    return 1; //access removed
+                } else{
+                    return 3; //nothing to remove
+                }
+            }else return 5; //cannot edit
+        }
+        else
+        {
+            return 2; //role not found
+        }
+    }
 
+    public int removeEdit(long processId, long whoEdits, User removeEdit){
+        Optional<Process> processData = processRepository.findById(processId);
+        if(processData.isPresent()) {
+            Process process_ = processData.get();
+            User whoEdits_ = userRepository.findById(whoEdits).get();
+            if(process_.getCanEdit().contains(whoEdits_)){
+                User removeEdit_ = userRepository.findById(removeEdit.getId()).get();
+                if(process_.getCanEdit().contains(removeEdit_)) {
+                    var list = process_.getCanEdit();
+                    list.remove(removeEdit_);
+                    process_.setCanEdit(list);
+                    processRepository.save(process_);
+                    return 1; //edit removed
+                } else{
+                    return 3; //nothing to remove
+                }
+            }else return 5; //cannot edit
+        }
+        else
+        {
+            return 2; //role not found
+        }
+    }
+
+    public int addEdit(long processId, long whoEdits, User getEdit){
+        Optional<Process> processData = processRepository.findById(processId);
+        if(processData.isPresent()) {
+            Process process_ = processData.get();
+            User whoEdits_ = userRepository.findById(whoEdits).get();
+            if(process_.getCanEdit().contains(whoEdits_)){
+                User getEdit_ = userRepository.findById(getEdit.getId()).get();
+                if(process_.getCanEdit().contains(getEdit_)){
+                    return 4; //already can edit
+                } else if(process_.getHasAccess().contains(getEdit_)) {
+                    var list = process_.getHasAccess();
+                    list.remove(getEdit_);
+                    process_.setHasAccess(list);
+                    list = process_.getCanEdit();
+                    list.add(getEdit_);
+                    process_.setCanEdit(list);
+                    processRepository.save(process_);
+                    for(Element e : process_.getElements())
+                    {
+                        if(e.getClass() == Task.class){
+                            taskService.addEdit(e.getId(), whoEdits_.getId(),getEdit);
+                        } else {
+                            this.addEdit(e.getId(),whoEdits_.getId(), getEdit); //
+                        }
+                    }
+                    return 1; //OK
+                } else{
+                    var list = process_.getCanEdit();
+                    list.add(getEdit_);
+                    process_.setCanEdit(list);
+                    processRepository.save(process_);
+                    return 1; //OK
+                }
+            }else return 5; //cannot edit
+        }
+        else
+        {
+            return 2; //role not found
+        }
+    }
+
+    public int deleteProcessById(long id, long whoEdits){
+        Optional<Process> processData = processRepository.findById(id);
+        if(processData.isPresent()) {
+            Process process_ = processData.get();
+            if (bpmnParser.removeProcessFromAllWorkflows(process_)) {
+                processRepository.deleteById(id);
+                return 1;
+            }
+            return 3;
+        }
+        return 2;
+    }
+
+    public int updateProcess(long id, Process process, long whoEdits){
+        Optional<Process> processData = processRepository.findById(id);
         if(processData.isPresent()){
             Process process_ = processData.get();
-            process_ = fillProcess(process_, process);
-            processRepository.save(process_);
-            bpmnParser.updateProcessInAllWorkflows(process_, true, null);
-            return 1;
+            User whoEdits_ = userRepository.findById(whoEdits).get();
+            if(process_.getCanEdit().contains(whoEdits_)) {
+                process_ = fillProcess(process_, process);
+                processRepository.save(process_);
+                bpmnParser.updateProcessInAllWorkflows(process_, true, null);
+                return 1;
+            }
+            return 3;
         }
         else
         {
@@ -117,6 +274,8 @@ public class ProcessService {
             }
             elementList.add(element_);
             process_.setElements(elementList);
+
+            //TODO add access and edit from process to element
 
             processRepository.save(process_);
             return 1;
@@ -138,7 +297,6 @@ public class ProcessService {
                 process_.setElements(elementList);
                 processRepository.save(process_);
                 return 1;
-
             }
             else {
                 return 3;
@@ -151,12 +309,16 @@ public class ProcessService {
         }
     }
 
-    public int saveWorkflow(long id, BPMNfile bpmn){
+    public int saveWorkflow(long id, BPMNfile bpmn, long whoEdits){
         Optional<Process> processData = processRepository.findById(id);
         if(processData.isPresent()) {
             Process process_ = processData.get();
-            bpmnParser.saveBPMN(bpmn, process_);
-            return 1;
+            User whoEdits_ = userRepository.findById(whoEdits).get();
+            if(process_.getCanEdit().contains(whoEdits_)) {
+                bpmnParser.saveBPMN(bpmn, process_);
+                return 1;
+            }
+            return 3;
         }
         else
         {
@@ -164,18 +326,22 @@ public class ProcessService {
         }
     }
 
-    public int restoreWorkflow(long id, HistoryBPMN bpmn){
+    public int restoreWorkflow(long id, HistoryBPMN bpmn, long whoEdits){
         if(processRepository.existsById(id)){
-            if(bpmnParser.canRestoreBPMN(bpmn.getBpmnContent())){
-                BPMNfile file = new BPMNfile();
-                Process process_ = processRepository.findById(id).get();
-                file.setProcess(process_);
-                file.setBpmnContent(bpmn.getBpmnContent());
-                bpmnParser.saveBPMN(file, process_);
-                return 1;
-            } else {
-                return 3;
+            Process process_ = processRepository.findById(id).get();
+            User whoEdits_ = userRepository.findById(whoEdits).get();
+            if(process_.getCanEdit().contains(whoEdits_)) {
+                if (bpmnParser.canRestoreBPMN(bpmn.getBpmnContent())) {
+                    BPMNfile file = new BPMNfile();
+                    file.setProcess(process_);
+                    file.setBpmnContent(bpmn.getBpmnContent());
+                    bpmnParser.saveBPMN(file, process_);
+                    return 1;
+                } else {
+                    return 4;
+                }
             }
+            return 3;
         } else {
             return 2;
         }
@@ -206,30 +372,37 @@ public class ProcessService {
         return 1;
     }
 
-    public List<Process> getAllTemplates(){
-        List<Process> allTemplates = processRepository.findAllTemplatesProcesses();
-        return allTemplates;
+    public List<Process> getAllTemplates(long userId){
+        if(userRepository.existsById(userId)) {
+            User user = userRepository.findById(userId).get();
+            List<Process> allTemplates = processRepository.findAllTemplatesProcessesForUser(user);
+             return allTemplates;
+        }else return null;
     }
 
-    public boolean addProcessFromFile(ProcessAndBpmnHolder holder){
+    public boolean addProcessFromFile(ProcessAndBpmnHolder holder, long whoEdits){
         Process newProcess = holder.getProcess();
         newProcess = processRepository.save(newProcess);
         BPMNfile newWorkflow = holder.getBpmn();
         newWorkflow.setBpmnContent(bpmnParser.prepareImportedFile(newWorkflow.getBpmnContent()));
 
-        this.saveWorkflow(newProcess.getId(), newWorkflow);
+        this.saveWorkflow(newProcess.getId(), newWorkflow, whoEdits);
         return true;
     }
 
-    public int addMetric(Long id, ProcessMetric metric) {
+    public int addMetric(Long id, ProcessMetric metric, long whoEdits) {
         Optional<Process> processData = processRepository.findById(id);
         if(processData.isPresent()) {
             Process process_ = processData.get();
+            User whoEdits_ = userRepository.findById(whoEdits).get();
+            if(process_.getCanEdit().contains(whoEdits_)) {
 
-            metric.setProcess(process_);
-            processMetricRepository.save(metric);
+                metric.setProcess(process_);
+                processMetricRepository.save(metric);
 
-            return 1;
+                return 1;
+            }
+            return 3;
         }
         else
         {
@@ -237,19 +410,21 @@ public class ProcessService {
         }
     }
 
-    public int removeMetric(Long id, ProcessMetric metric) {
+    public int removeMetric(Long id, ProcessMetric metric, long whoEdits) {
         Optional<Process> processData = processRepository.findById(id);
         if(processData.isPresent()) {
             Process process_ = processData.get();
-            ProcessMetric metric_ = processMetricRepository.findById(metric.getId()).get();
-            if(metric_.getProcess().getId() == process_.getId())
-            {
-                processMetricRepository.delete(metric_);
-                return 1;
+            User whoEdits_ = userRepository.findById(whoEdits).get();
+            if(process_.getCanEdit().contains(whoEdits_)) {
+                ProcessMetric metric_ = processMetricRepository.findById(metric.getId()).get();
+                if (metric_.getProcess().getId() == process_.getId()) {
+                    processMetricRepository.delete(metric_);
+                    return 1;
+                } else {
+                    return 4;
+                }
             }
-            else {
-                return 3;
-            }
+            return 3;
 
         }
         else
